Move Draft row out to ListView story and component
