+++ packages/ui/src/Integration/IntegrationDetailHistoryListView.tsx
@@ -1,7 +1,10 @@
-import { Grid } from 'patternfly-react';
+import { Button, Grid, ListView, ListViewItem } from 'patternfly-react';
 import * as React from 'react';
 
 export interface IIntegrationDetailHistoryListViewProps {
+  integrationIsDraft: boolean;
+  i18nTextBtnEdit?: string;
+  i18nTextBtnPublish?: string;
   i18nTextDraft?: string;
   i18nTextHistory?: string;
 }
@@ -13,23 +16,39 @@ export class IntegrationDetailHistoryListView extends React.Component<
     return (
       <>
         <Grid fluid={true} key={1}>
-          <Grid.Row className="show-grid">
-            <Grid.Col xs={2} md={2}>
-              {<span>{this.props.i18nTextDraft}:</span>}
-            </Grid.Col>
-            <Grid.Col xs={10} md={10}>
-              <p>blah</p>
-            </Grid.Col>
-          </Grid.Row>
+          {this.props.integrationIsDraft ? (
+            <Grid.Row className="show-grid">
+              <Grid.Col xs={2} md={2}>
+                {<span>{this.props.i18nTextDraft}:</span>}
+              </Grid.Col>
+              <Grid.Col xs={10} md={10}>
+                <ListViewItem
+                  key={1}
+                  heading={this.props.i18nTextDraft}
+                  actions={
+                    <>
+                      <Button>{this.props.i18nTextBtnPublish}</Button>
+                      <Button>{this.props.i18nTextBtnEdit}</Button>
+                    </>
+                  }
+                  stacked={false}
+                />
+              </Grid.Col>
+            </Grid.Row>
+          ) : null}
 
-          <Grid.Row className="show-grid">
-            <Grid.Col xs={2} md={2}>
-              {<span>{this.props.i18nTextHistory}:</span>}
-            </Grid.Col>
-            <Grid.Col xs={10} md={10}>
-              <p>blah</p>
-            </Grid.Col>
-          </Grid.Row>
+          {this.props.children ? (
+            <Grid.Row className="show-grid">
+              <Grid.Col xs={2} md={2}>
+                {<span>{this.props.i18nTextHistory}:</span>}
+              </Grid.Col>
+              <Grid.Col xs={10} md={10}>
+                {this.props.children ? (
+                  <ListView>{this.props.children}</ListView>
+                ) : null}
+              </Grid.Col>
+            </Grid.Row>
+          ) : null}
         </Grid>
       </>
     );
