+++ packages/ui/stories/Integration/IntegrationDetailHistoryListView.stories.tsx
@@ -10,15 +10,51 @@ const stories = storiesOf(
   module
 );
 
+const integrationPublished = {
+  id: 'i-LYmlhVFB6pKKaBQVSyez',
+  version: 1,
+  updatedAt: 1550261344272,
+  currentState: 'Unpublished',
+  targetState: 'Unpublished',
+  name: 'aaa',
+  isDraft: false,
+};
+
+const integrationUnpublished = {
+  id: 'i-LYmlhVFB6pKKaBQVSyez',
+  version: 1,
+  updatedAt: 1550261344272,
+  currentState: 'Unpublished',
+  targetState: 'Unpublished',
+  name: 'aaa',
+  isDraft: true,
+};
+
+const i18nTextBtnEdit = 'Edit';
+const i18nTextBtnPublish = 'Publish';
 const i18nTextDraft = 'Draft';
 const i18nTextHistory = 'History';
 
-stories.add(
-  'has history',
-  withNotes('Verify there is a list of history items')(() => (
-    <IntegrationDetailHistoryListView
-      i18nTextDraft={text('i18nTextDraft', i18nTextDraft)}
-      i18nTextHistory={text('i18nTextHistory', i18nTextHistory)}
-    />
-  ))
-);
+stories
+  .add(
+    'published',
+    withNotes('Verify there is no Publish button')(() => (
+      <IntegrationDetailHistoryListView
+        integrationIsDraft={integrationPublished.isDraft}
+        i18nTextDraft={text('i18nTextDraft', i18nTextDraft)}
+        i18nTextHistory={text('i18nTextHistory', i18nTextHistory)}
+      />
+    ))
+  )
+  .add(
+    'draft, no published history',
+    withNotes('Verify there is a Publish button')(() => (
+      <IntegrationDetailHistoryListView
+        integrationIsDraft={integrationUnpublished.isDraft}
+        i18nTextBtnEdit={text('i18nTextBtnEdit', i18nTextBtnEdit)}
+        i18nTextBtnPublish={text('i18nTextBtnPublish', i18nTextBtnPublish)}
+        i18nTextDraft={text('i18nTextDraft', i18nTextDraft)}
+        i18nTextHistory={text('i18nTextHistory', i18nTextHistory)}
+      />
+    ))
+  );
